#1701: extract security-policy types into types_security.go
Moves the security domain type definitions verbatim out of types.go into
types_security.go (same package config):

  SecurityConfig, zones (ZoneConfig, HostInboundTraffic), policy
  (ZonePairPolicies, Policy, PolicyMatch, PolicyAction+const, PolicyLog),
  flow/session (FlowConfig, FlowTraceoptions, TracePacketFilter,
  TCPSessionConfig), ALGConfig, logging (LogConfig, SyslogTransport,
  SyslogStream, SSHKnownHostKey, PreIDDefaultPolicy), all NAT
  (NATConfig..StaticNATRule, NATType+const), all Screen/IDS
  (LimitSessionScreen..SynFloodConfig), address book (AddressBook,
  Address, AddressSet), applications (ApplicationsConfig, ApplicationSet,
  Application), IPsec/IKE (IPsecConfig..IPsecVPN), dynamic-address feeds
  (DynamicAddressConfig, FeedServer, FeedEntry, AddressBinding), and the
  [edit schedulers] time-range SchedulerConfig.

IPsec/IKE, DynamicAddress/Feed, and SchedulerConfig are placed here per
the plan's AGY/Codex re-bucket: all three nest under SecurityConfig or a
security policy reference (Policy.SchedulerName), distinct from the
similarly-named CoSScheduler. The PolicyAction and NATType iota const
blocks move as intact units with their owning types so ordinals are
unchanged. Pure byte-identical motion; consumers untouched. Verified: go
build ./... clean, vet clean, tests pass.
